feat(db): ajustar modelo a la base de datos.

+++ src/main/java/ar/edu/utn/frc/siga/allocation/dto/response/AllocationHistorySnapshotDto.java
@@ -7,7 +7,7 @@ import java.time.Instant;
 public record AllocationHistorySnapshotDto(
         Long id,
         Long occurrenceId,
-        Integer classroomId,
+        Long classroomId,
         AllocationSource source,
         Instant createdAt,
         String observation
