+++ src/main/java/ar/edu/utn/frc/siga/space/dto/request/BuildingActiveBatchItemDto.java
@@ -3,7 +3,7 @@ package ar.edu.utn.frc.siga.space.dto.request;
 import jakarta.validation.constraints.NotNull;
 
 public record BuildingActiveBatchItemDto(
-    @NotNull Integer id,
+    @NotNull Long id,
     @NotNull Boolean active
 ) {
 }
