+++ src/main/java/ar/edu/utn/frc/siga/space/repository/BuildingRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BuildingRepository extends JpaRepository<Building, Integer> {
+public interface BuildingRepository extends JpaRepository<Building, Long> {
 
     Optional<Building> findByName(String name);
