+++ src/main/java/ar/edu/utn/frc/siga/space/repository/ClassroomTypeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ClassroomTypeRepository extends JpaRepository<ClassroomType, Integer> {
+public interface ClassroomTypeRepository extends JpaRepository<ClassroomType, Long> {
 
     Optional<ClassroomType> findByDescriptionIgnoreCase(String description);
